ssh vnc 全部交给supervisor管理

+++ ubuntu/src/start_up.sh
@@ -2,6 +2,9 @@
 ### every exit != 0 fails the script
 set -e
 
+ENABLE_SSH=${ENABLE_SSH:-true}
+ENABLE_VNC=${ENABLE_VNC:-true}
+
 ## fix network config
 if [ -d "/root/dockerstartup/networkbackup/" ]; then
     cat /root/dockerstartup/networkbackup/hosts > /etc/hosts
@@ -9,28 +12,31 @@ if [ -d "/root/dockerstartup/networkbackup/" ]; then
     cat /root/dockerstartup/networkbackup/resolv.conf > /etc/resolv.conf
 fi
 
-## change vnc password
-echo -e "\n------------------ change VNC password  ------------------"
-# first entry is control, second is view (if only one is valid for both)
-mkdir -p "$HOME/.vnc"
-PASSWD_PATH="$HOME/.vnc/passwd"
+SSH_CONF=/etc/supervisor/conf.d/sshd.conf
+VNC_CONF=/etc/supervisor/conf.d/vncserver.conf
 
-if [[ -f $PASSWD_PATH ]]; then
-    echo -e "\n---------  purging existing VNC password settings  ---------"
-    rm -f $PASSWD_PATH
+# start ssh
+if [[ "$ENABLE_SSH" == "true" ]] ; then
+    if [ ! -f $SSH_CONF ] ; then
+        cp /root/dockerstartup/modes/sshd.conf /etc/supervisor/conf.d/
+    fi
 fi
 
-echo "$VNC_PW" | vncpasswd -f >> $PASSWD_PATH
-chmod 600 $PASSWD_PATH
+# start vnc server
+if [[ "$ENABLE_VNC" == "true" ]] ; then
+    if [ ! -f $VNC_CONF ] ; then
+        echo "remove old vnc locks to be a reattachable container"
+        vncserver -kill :1 \
+            || rm -rfv /tmp/.X*-lock /tmp/.X11-unix \
+            || echo "no locks present"
 
-echo -e "\n------------------ start VNC server ------------------------"
-echo "remove old vnc locks to be a reattachable container"
-vncserver -kill :1 \
-    || rm -rfv /tmp/.X*-lock /tmp/.X11-unix \
-    || echo "no locks present"
-vncserver -geometry ${WIDTH:-1152}x${HEIGHT-864} :1 securitytypes=none
+        cp /root/dockerstartup/modes/vncserver.conf /etc/supervisor/conf.d/
+        sed -i "s/1152/${WIDTH:-1152}/g" $VNC_CONF
+        sed -i "s/864/${HEIGHT:-864}/g" $VNC_CONF
+    fi
+fi
 
-#start supervisor
+# Forward SIGTERM to supervisord process
 _term() {
     while kill -0 $child >/dev/null 2>&1
     do
